week_9/ai-agents-mini-bootcamp-main/RAG/requirements.txt
langchain
langchain-ollama
langchain-chroma
pandas
openai
python-dotenv